requirements.txt
google-genai
pyaudio
python-dotenv
websockets
